Move developer-focused commands to dev-cmd.
