+++ Library/Homebrew/dev-cmd/man.rb
